Prepare bower

+++ prepare_bower.sh
@@ -3,7 +3,11 @@
 echo "Prepare Bower"
 
 if [ -z "$CIRCLECI" ]; then
-  rm -rf ./vendor/assets/bower_components
+
+  if [ -d ./vendor/assets/bower_components ]; then
+    echo "Exists remove..."
+    rm -rf ./vendor/assets/bower_components
+  fi
 fi
 
 ./node_modules/bower/bin/bower install
